refactor: moved subtasks from task dir
